requirements.txt
numpy       
Pillow
